requirements/local.txt
-r base.txt

coverage==3.7.1
django-debug-toolbar==1.3.0
ipython==3.1.0
py==1.4.26
pytest==2.7.0
pytest-django==2.8.0
sqlparse==0.1.14
